move major combobox
